requirements.txt
Django==5.2.6
djangorestframework==3.16.1
python-decouple==3.8
django-cors-headers==4.7.0
web3==6.15.1
requests==2.31.0
cryptography==42.0.5
